requirements.txt
acme==3.0.1
aiodns==3.2.0
aiohappyeyeballs==2.6.1
aiohasupervisor==0.2.2b5
aiohttp==3.11.11
aiohttp-client-cache==0.14.2
aiohttp-cors==0.7.0
aiohttp-fast-zlib==0.2.0
aiooui==0.1.9
aioredis==2.0.1
aiosignal==1.4.0
aiozoneinfo==0.2.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
aplay==1.1.1
astral==2.2
async-timeout==5.0.1
async_interrupt==1.2.0
atomicwrites-homeassistant==1.4.1
attmap==0.13.2
attrs==25.4.0
av==16.0.1
awesomeversion==24.6.0
bcrypt==4.2.0
bleak==2.1.0
bleak-retry-connector==4.4.4
bluetooth-adapters==2.1.1
bluetooth-auto-recovery==1.5.3
bluetooth-data-tools==1.28.4
boto3==1.42.17
botocore==1.42.17
btsocket==0.3.0
cattrs==25.3.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
ciso8601==2.3.2
click==8.3.1
coloredlogs==15.0.1
cronsim==2.6
cryptography==44.0.0
ctranslate2==4.6.2
dbus-fast==3.1.2
dnspython==2.8.0
eido==0.2.4
email-validator==2.3.0
envs==1.4
fastapi==0.123.9
fastapi-cli==0.0.16
fastapi-cloud-cli==0.6.0
fastar==0.8.0
faster-whisper==1.2.1
filelock==3.20.0
flatbuffers==25.9.23
fnv-hash-fast==1.0.2
fnvhash==0.1.0
frozenlist==1.8.0
fsspec==2025.12.0
greenlet==3.3.0
h11==0.16.0
habluetooth==5.8.0
hass-nabucasa==0.87.0
hf-xet==1.2.0
home-assistant-bluetooth==1.13.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.2
huggingface_hub==1.2.3
humanfriendly==10.0
idna==3.11
ifaddr==0.2.0
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.5.2
josepy==1.15.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
llvmlite==0.46.0
logmuse==0.2.8
lru-dict==1.3.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mashumaro==3.17
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
networkx==3.6.1
numba==0.63.1
numpy==2.3.5
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
onnxruntime==1.23.2
openwakeword==0.4.0
orjson==3.10.12
oyaml==1.0
packaging==25.0
pandas==2.3.3
pephubclient==0.4.5
peppy==0.40.8
pillow==11.0.0
piper==0.14.5
piper-tts==1.3.0
pipestat==0.12.2
platformdirs==4.5.1
propcache==0.2.1
protobuf==6.33.2
psutil==7.1.3
psutil-home-assistant==0.0.1
pycares==5.0.0
pycognito==2024.5.1
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==24.3.0
pyRFC3339==2.1.0
PyRIC==0.1.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
referencing==0.37.0
regex==2025.11.3
requests==2.32.3
requests-cache==1.2.1
rich==14.2.0
rich-toolkit==0.17.0
rignore==0.7.6
rpds-py==0.30.0
s3transfer==0.16.0
scikit-learn==1.8.0
scipy==1.16.3
securetar==2024.11.0
sentry-sdk==2.47.0
setuptools==80.9.0
shellingham==1.5.4
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
snitun==0.39.1
sounddevice==0.5.3
SQLAlchemy==2.0.36
starlette==0.50.0
stua==0.3
sympy==1.14.0
text-unidecode==1.3
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
triton==3.5.1
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uart-devices==0.1.1
ubiquerg==0.8.2
ulid-transform==1.0.2
url-normalize==2.2.1
urllib3==1.26.20
usb-devices==0.4.5
uv==0.5.8
uvicorn==0.38.0
uvloop==0.22.1
voluptuous==0.15.2
voluptuous-openapi==0.0.5
voluptuous-serialize==2.6.0
watchfiles==1.1.1
webrtc-models==0.3.0
websockets==15.0.1
yacman==0.9.4
yarl==1.18.3
